LatticeModels.jl

LatticeModels.Basis

LatticeModels.Bravais

LatticeModels.HoneycombLattice

LatticeModels.Lattice

LatticeModels.LatticeArray

LatticeModels.LatticeOperator

LatticeModels.LatticeSite

LatticeModels.LatticeValue

LatticeModels.SquareLattice

LatticeModels.TensorProduct

LatticeModels.coord_operators

LatticeModels.coord_operators

LatticeModels.coord_values

LatticeModels.coords

LatticeModels.diag_aggregate

LatticeModels.diag_operator

LatticeModels.diag_operator

LatticeModels.diag_operator

LatticeModels.diag_operator

LatticeModels.evolution_operator

LatticeModels.plot_fallback

LatticeModels.radius_vector

LatticeModels.sublattice

LatticeModels.@evolution

LatticeModels.@on_lattice

Lattice creation

LatticeModels.Bravais — Type

Bravais{N, NB}

N-dimensional infinite Bravais lattice with NB sites in basis.

Bravais(translation_vectors[, basis])

Constructs a Bravais lattice with given translation vectors and locations of basis sites relative to some unit cell. The basis argument can be omitted, in which case the lattice basis will consist of one site located in the bottom-left corner of the unit cell.

translation_vectors argument must be an AbstractMatrix{<:Real} of size N×N, while basis must also be an abstract matrix of size N×NB.

LatticeModels.HoneycombLattice — Type

HoneycombLattice

Type alias for Lattice{:honeycomb,2,2}.

HoneycombLattice(sz::Vararg{Int, 2})

Constructs a honeycomb lattice with a sz-size macro cell.

LatticeModels.Lattice — Type

Lattice{LatticeSym, N, NB}

A finite subset of a Brvais{N, NB}. LatticeSym is a Symbol which represents the type of the lattice (e. g. :square, :honeycomb). This makes Lattice object behavior known at compile-time, which allows to introduce various optimizations or to define specific plot recipes.

Lattice(sym, sz, bvs[, mask])

Constructs a finite Lattice{sym, N, NB} as a subset of the bvs Bravais lattice. sz is a NTuple{N, Int} which represents how many times the unit cell of bvs was translated by each axis - these sites form a macro cell. mask, if defined, is a Vector{Bool} storing information about which of the sites from the macro cell are actually included in the lattice, and which are not.

For example, a 3×3 square lattice with its center site excluded is represented as Lattice(:square, (3, 3), Bravais([1 0; 0 1]), Bool[1, 1, 1, 1, 0, 1, 1, 1, 1])

To define a new type of lattice, create an alias for Lattice{YourSym, YourN, YourNB}. Refer to the docs for detailed explanation.

LatticeModels.LatticeSite — Type

LatticeSite{N}

A site of a Lattice{LatticeSym, N, NB} lattice.

Fields:

unit_cell: a set of translations along all axes representing the unit cell the site is located in.

basis_index: the number of site in the lattice basis.

This type is used to iterate over all sites of a Lattice{LatticeSym, N, NB}. The exact location of a LatticeSite can be found using the coords(lattice, site) function.

LatticeModels.SquareLattice — Type

SquareLattice{N}

Type alias for Lattice{:square,N,1}.

SquareLattice(sz::Int...)

Constructs a square lattice of size sz.

LatticeModels.coords — Method

coords(lattice::Lattice, site::LatticeSite) -> vector

Finds the location in space of lattice site site on lattice lattice.

LatticeModels.radius_vector — Method

radius_vector(lattice::Lattice, site1::LatticeSite, site2::LatticeSite) -> vector

Finds the vector between two sites on a lattice according to possibly periodic boundary conditions (site2 will be translated along the macro cell to minimize the distance between them).

LatticeModels.sublattice — Method

sublattice(lf::Function, l::Lattice) -> Lattice

Generates a a subset of lattice l by applying the lf function to its sites. The lf function must accept two positional arguments (a LatticeSite and a vector with its coordinates) and return a boolean value.

Lattice values

LatticeModels.LatticeValue — Type

LatticeValue{T, LatticeSym}

Represents a value of type T on a Lattice{LatticeSym} lattice.

Fields:

lattice: the Lattice object the value is defined on

values: the values on different sites

LatticeModels.coord_values — Method

coord_values(lattice::Lattice)

Generates a tuple of LatticeValues representing coordinate functions.

LatticeModels.plot_fallback — Method

plot_fallback(lv::LatticeValue)

Creates a copy of lv lattice value with its LatticeSym overwritten to :uncertain. Use it to invoke the default plot recipe for LatticeValues when defining a custom one.

Lattice operators

LatticeModels.Basis — Type

Basis{LT} where {LT<:Lattice}

A basis on a lattice with some number of internal states on each site. Fields:

lattice: the Lattice of the basis

internal_dim: the number of internal states on each site

LatticeModels.LatticeArray — Type

LatticeArray{LT, MT} where {LT<:Lattice, MT<:AbstractArray}

A wrapper object for array representing a wave function or linear operator. Stores information about its basis to perform lattice checks.

LatticeModels.LatticeOperator — Method

LatticeOperator{LT, MT}

The same as LatticeArray{LT, MT} where MT<:AbstractMatrix.

LatticeOperator(uniform_scaling, basis)

Creates a LatticeOperator representation of a UniformScaling operator on given basis.

LatticeModels.TensorProduct — Type

TensorProduct{LVT, MT} where {LVT<:LatticeValue{<:Number}, MT<:AbstractMatrix}

A lazy representation of an operator as a tensor product of two distinct phase spaces. One affects only the internal space, the other - only the lattice space.

The lattice_value ⊗ matrix notation computes the value of the TensorProduct eagerly, which means that the result will be a LatticeOperator. However, in the @hamiltonian macro lazy computation is forced.

LatticeModels.coord_operators — Method

coord_operators(basis)

Returns a Tuple of coordinate LatticeOperators for given basis.

LatticeModels.coord_operators — Method

coord_operators(lattice, ndims)

The same as coord_operators(Basis(lattice, ndims)).

LatticeModels.diag_aggregate — Method

diag_aggregate(function, lattice_operator)

Creates a LatticeValue where a site maps to the result of function on the matrix of the operator narrowed to that site.

LatticeModels.diag_operator — Method

diag_operator(basis, lattice_value)

Creates a diagonal operator which affects only the lattice space. The lattice_value argument must be a LatticeValue storing diagonal elements of the operator in lattice space.

LatticeModels.diag_operator — Method

diag_operator(site_fun, basis)

Creates a diagonal operator which affects only the lattice space. The site_fun function must accept a LatticeSite and its coordinates and return a number which will be the diagonal element of the operator in lattice space.

LatticeModels.diag_operator — Method

diag_operator(site_fun, lattice)

Creates a diagonal operator by applying the site_fun function to each site of lattice. The site_fun function must accept a LatticeSite and its coordinates and return a matrix which will be an operator affecting the internal state of the site.

LatticeModels.diag_operator — Method

diag_operator(lattice, matrix)

Creates a diagonal operator which affects only the internal state the same way on every site. matrix is an AbstractMatrix representing the linear operator on the internal space.

Note that the matrix of the output LatticeOperator will be similar to matrix: for instance, if matrix is sparse, so will be the output.

LatticeModels.@on_lattice — Macro

@on_lattice

Replaces all LatticeArrays in subsequent function calls with actual arrays stored inside them. Throws an error if lattice operators in one function call are defined on different lattices, shows a warning if a lattice array is used in one call with a normal array.

Example

l = SquareLattice(10, 10)
bas = Basis(l, 2)
X, Y = coord_operators(bas)
xexpypy = diag_operator(bas) do site, (x, y)
    x * exp(y) + y
end
xexpy == @on_lattice X * exp(Y) + Y     # true

Hoppings

Magnetic fields

Hamiltonian tools

Currents

Unitary evolution

LatticeModels.evolution_operator — Method

evolution_operator(H, t[, k])

Calculates the unitary evolution operator using the formula

$ \mathcal{U}(t) = e^{-\frac{1}{i\hbar} \hat{H} t} $

Arguments

H: the hamiltonian matrix

t: the evolution time

k: if provided, the exponent will be calculated using a Taylor series expansion with order k

LatticeModels.@evolution — Macro

@evolution [kwargs...] {rules...} for_loop

Generates an environment with defined hamiltonian and density matrices that evolve by certain laws. See Unitary evolution for more details.