LatticeModels.jl

LatticeModels.Bravais

LatticeModels.HoneycombLattice

LatticeModels.Lattice

LatticeModels.LatticeSite

LatticeModels.SquareLattice

LatticeModels.coords

LatticeModels.evolution_operator

LatticeModels.radius_vector

LatticeModels.sublattice

LatticeModels.@evolution

Lattice creation

LatticeModels.Bravais — Type

Bravais{N, NB}

N-dimensional infinite Bravais lattice with NB sites in basis.

Bravais(translation_vectors[, basis])

Constructs a Bravais lattice with given translation vectors and locations of basis sites relative to some unit cell. The basis argument can be omitted, in which case the lattice basis will consist of one site located in the bottom-left corner of the unit cell.

translation_vectors argument must be an AbstractMatrix{<:Real} of size N×N, while basis must also be an abstract matrix of size N×NB.

LatticeModels.HoneycombLattice — Type

HoneycombLattice

Type alias for Lattice{:honeycomb,2,2}.

HoneycombLattice(xsz::Int, ysz::Int)

Constructs a honeycomb lattice with a xsz×ysz macro cell.

LatticeModels.Lattice — Type

Lattice{LatticeSym, N, NB}

A finite subset of a Brvais{N, NB}. LatticeSym is a Symbol which represents the type of the lattice (e. g. :square, :honeycomb). This makes Lattice object behavior known at compile-time, which allows to introduce various optimizations or to define specific plot recipes.

Lattice(sym, sz, bvs[, mask])

Constructs a finite Lattice{sym, N, NB} as a subset of the bvs Bravais lattice. sz is a NTuple{N, Int} which represents how many times the unit cell of bvs was translated by each axis - these sites form a macro cell. mask, if defined, is a Vector{Bool} storing information about which of the sites from the macro cell are actually included in the lattice, and which are not.

For example, a 3×3 square lattice with its center site excluded is represented as Lattice(:square, (3, 3), Bravais([1 0; 0 1]), Bool[1, 1, 1, 1, 0, 1, 1, 1, 1])

To define a new type of lattice, create an alias for Lattice{YourSym, YourN, YourNB}. Refer to the docs for detailed explanation.

LatticeModels.LatticeSite — Type

LatticeSite{N}

A site of a Lattice{LatticeSym, N, NB} lattice. Fields:

unit_cell: a set of translations along all axes representing the unit cell the site is located in.

basis_index: the number of site in the lattice basis.

This type is used to iterate over all sites of a Lattice{LatticeSym, N, NB}. The exact location of a LatticeSite can be found using the coords(lattice, site) function.

LatticeModels.SquareLattice — Type

SquareLattice{N}

Type alias for Lattice{:square,N,1}.

SquareLattice(sz::Int...)

Constructs a square lattice of size sz.

LatticeModels.coords — Method

coords(lattice::Lattice, site::LatticeSite) -> vector

Finds the location in space of lattice site site on lattice lattice.

LatticeModels.radius_vector — Method

radius_vector(lattice::Lattice, site1::LatticeSite, site2::LatticeSite) -> vector

Finds the vector between two sites on a lattice according to possibly periodic boundary conditions (site2 will be translated along the macro cell to minimize the distance between them).

LatticeModels.sublattice — Method

sublattice(lf::Function, l::Lattice) -> Lattice

Generates a a subset of lattice l by applying the lf function to its sites. The lf function must accept two positional arguments (a LatticeSite and a vector with its coordinates) and return a boolean value.

Lattice values

Lattice operators

Hoppings

Magnetic fields

Hamiltonian tools

Currents

Unitary evolution

LatticeModels.evolution_operator — Method

evolution_operator(H, t[, k])

Calculates the unitary evolution operator using the formula

$ \mathcal{U}(t) = e^{-\frac{1}{i\hbar} \hat{H} t} $

Arguments

H: the hamiltonian matrix

t: the evolution time

k: if provided, the exponent will be calculated using a Taylor series expansion with order k

LatticeModels.@evolution — Macro

@evolution [kwargs...] {rules...} for_loop

Generates an environment with defined hamiltonian and density matrices that evolve by certain laws. See Unitary evolution for more details.