LatticeModels.jl

LatticeModels.AbstractCurrents

LatticeModels.Basis

LatticeModels.BondSet

LatticeModels.Bravais

LatticeModels.DensityCurrents

LatticeModels.FluxField

LatticeModels.HoneycombLattice

LatticeModels.Hopping

LatticeModels.LandauField

LatticeModels.Lattice

LatticeModels.LatticeArray

LatticeModels.LatticeOperator

LatticeModels.LatticeSite

LatticeModels.LatticeValue

LatticeModels.MaterializedCurrents

LatticeModels.NoField

LatticeModels.Spectrum

LatticeModels.SquareLattice

LatticeModels.SubCurrents

LatticeModels.SymmetricField

LatticeModels.TensorProduct

LatticeModels.apply_field!

LatticeModels.bonds

LatticeModels.bonds

LatticeModels.coord_operators

LatticeModels.coord_operators

LatticeModels.coord_values

LatticeModels.coords

LatticeModels.current_lambda

LatticeModels.diag_aggregate

LatticeModels.diag_operator

LatticeModels.diag_operator

LatticeModels.diag_operator

LatticeModels.diag_operator

LatticeModels.evolution_operator

LatticeModels.filled_projector

LatticeModels.hopping

LatticeModels.hopping_operator

LatticeModels.lattice

LatticeModels.materialize

LatticeModels.plot_fallback

LatticeModels.projector

LatticeModels.projector

LatticeModels.promote_dims!

LatticeModels.ptrace

LatticeModels.radius_vector

LatticeModels.spectrum

LatticeModels.sublattice

LatticeModels.trip_integral

LatticeModels.vector_potential

LatticeModels.@evolution

LatticeModels.@field_def

LatticeModels.@hamiltonian

LatticeModels.@on_lattice

Lattice creation

LatticeModels.Bravais — Type

Bravais{N, NB}

N-dimensional infinite Bravais lattice with NB sites in basis.

Bravais(translation_vectors[, basis])

Constructs a Bravais lattice with given translation vectors and locations of basis sites relative to some unit cell. The basis argument can be omitted, in which case the lattice basis will consist of one site located in the bottom-left corner of the unit cell.

translation_vectors argument must be an AbstractMatrix{<:Real} of size N×N, while basis must also be an abstract matrix of size N×NB.

LatticeModels.HoneycombLattice — Type

HoneycombLattice

Type alias for Lattice{:honeycomb,2,2}.

HoneycombLattice(sz::Vararg{Int, 2})

Constructs a honeycomb lattice with a sz-size macro cell.

LatticeModels.Lattice — Type

Lattice{LatticeSym, N, NB}

A finite subset of a Brvais{N, NB}. LatticeSym is a Symbol which represents the type of the lattice (e. g. :square, :honeycomb). This makes Lattice object behavior known at compile-time, which allows to introduce various optimizations or to define specific plot recipes.

Lattice(sym, sz, bvs[, mask])

Constructs a finite Lattice{sym, N, NB} as a subset of the bvs Bravais lattice. sz is a NTuple{N, Int} which represents how many times the unit cell of bvs was translated by each axis - these sites form a macro cell. mask, if defined, is a Vector{Bool} storing information about which of the sites from the macro cell are actually included in the lattice, and which are not.

For example, a 3×3 square lattice with its center site excluded is represented as Lattice(:square, (3, 3), Bravais([1 0; 0 1]), Bool[1, 1, 1, 1, 0, 1, 1, 1, 1])

To define a new type of lattice, create an alias for Lattice{YourSym, YourN, YourNB}. Refer to the docs for detailed explanation.

LatticeModels.LatticeSite — Type

LatticeSite{N}

A site of a Lattice{LatticeSym, N, NB} lattice.

Fields:

unit_cell: a set of translations along all axes representing the unit cell the site is located in.

basis_index: the number of site in the lattice basis.

This type is used to iterate over all sites of a Lattice{LatticeSym, N, NB}. The exact location of a LatticeSite can be found using the coords(lattice, site) function.

LatticeModels.SquareLattice — Type

SquareLattice{N}

Type alias for Lattice{:square,N,1}.

SquareLattice(sz::Int...)

Constructs a square lattice of size sz.

LatticeModels.coords — Method

coords(lattice::Lattice, site::LatticeSite) -> vector

Finds the location in space of lattice site site on lattice lattice.

LatticeModels.radius_vector — Method

radius_vector(lattice::Lattice, site1::LatticeSite, site2::LatticeSite) -> vector

Finds the vector between two sites on a lattice according to possibly periodic boundary conditions (site2 will be translated along the macro cell to minimize the distance between them).

LatticeModels.sublattice — Method

sublattice(lf::Function, l::Lattice) -> Lattice

Generates a a subset of lattice l by applying the lf function to its sites. The lf function must accept two positional arguments (a LatticeSite and a vector with its coordinates) and return a boolean value.

Lattice values

LatticeModels.LatticeValue — Type

LatticeValue{T, LatticeSym}

Represents a value of type T on a Lattice{LatticeSym} lattice.

Fields:

lattice: the Lattice object the value is defined on

values: the values on different sites

LatticeModels.coord_values — Method

coord_values(lattice::Lattice)

Generates a tuple of LatticeValues representing coordinate functions.

LatticeModels.plot_fallback — Method

plot_fallback(lv::LatticeValue)

Creates a copy of lv lattice value with its LatticeSym overwritten to :uncertain. Use it to invoke the default plot recipe for LatticeValues when defining a custom one.

Lattice operators

LatticeModels.Basis — Type

Basis{LT} where {LT<:Lattice}

A basis on a lattice with some number of internal states on each site. Fields:

lattice: the Lattice of the basis

internal_dim: the number of internal states on each site

LatticeModels.LatticeArray — Type

LatticeArray{LT, MT} where {LT<:Lattice, MT<:AbstractArray}

A wrapper object for array representing a wave function or linear operator. Stores information about its basis to perform lattice checks.

LatticeModels.LatticeOperator — Method

LatticeOperator{LT, MT}

The same as LatticeArray{LT, MT} where MT<:AbstractMatrix.

LatticeOperator(uniform_scaling, basis)

Creates a LatticeOperator representation of a UniformScaling operator on given basis.

LatticeModels.TensorProduct — Type

TensorProduct{LVT, MT} where {LVT<:LatticeValue{<:Number}, MT<:AbstractMatrix}

A lazy representation of an operator as a tensor product of two distinct phase spaces. One affects only the internal space, the other - only the lattice space.

The lattice_value ⊗ matrix notation computes the value of the TensorProduct eagerly, which means that the result will be a LatticeOperator. However, in the @hamiltonian macro lazy computation is forced.

LatticeModels.coord_operators — Method

coord_operators(basis)

Returns a Tuple of coordinate LatticeOperators for given basis.

LatticeModels.coord_operators — Method

coord_operators(lattice, ndims)

The same as coord_operators(Basis(lattice, ndims)).

LatticeModels.diag_aggregate — Method

diag_aggregate(function, lattice_operator)

Creates a LatticeValue where a site maps to the result of function on the matrix of the operator narrowed to that site.

LatticeModels.diag_operator — Method

diag_operator(basis, lattice_value)

Creates a diagonal operator which affects only the lattice space. The lattice_value argument must be a LatticeValue storing diagonal elements of the operator in lattice space.

LatticeModels.diag_operator — Method

diag_operator(site_fun, basis)

Creates a diagonal operator which affects only the lattice space. The site_fun function must accept a LatticeSite and its coordinates and return a number which will be the diagonal element of the operator in lattice space.

LatticeModels.diag_operator — Method

diag_operator(site_fun, lattice)

Creates a diagonal operator by applying the site_fun function to each site of lattice. The site_fun function must accept a LatticeSite and its coordinates and return a matrix which will be an operator affecting the internal state of the site.

LatticeModels.diag_operator — Method

diag_operator(lattice, matrix)

Creates a diagonal operator which affects only the internal state the same way on every site. matrix is an AbstractMatrix representing the linear operator on the internal space.

Note that the matrix of the output LatticeOperator will be similar to matrix: for instance, if matrix is sparse, so will be the output.

LatticeModels.ptrace — Method

ptrace(lattice_operator)

Same as diag_aggregate(tr, lattice_operator)

LatticeModels.@on_lattice — Macro

@on_lattice

Replaces all LatticeArrays in subsequent function calls with actual arrays stored inside them. Throws an error if lattice operators in one function call are defined on different lattices, shows a warning if a lattice array is used in one call with a normal array.

Example

l = SquareLattice(10, 10)
bas = Basis(l, 2)
X, Y = coord_operators(bas)
xexpypy = diag_operator(bas) do site, (x, y)
    x * exp(y) + y
end
xexpy == @on_lattice X * exp(Y) + Y     # true

Hoppings

LatticeModels.BondSet — Type

BondSet{LT} where {LT<:Lattice}

Represents the bonds on some lattice.

BondSets can be combined with the | operator and negated with the ! operator. Also you can create a BondSet which connects sites that were connected by ≤n bonds of the previous BondSet by taking its power: bs2 = bs1 ^ n.

LatticeModels.Hopping — Type

Hopping{MT} where {MT<:AbstractMatrix}

A struct representing a bond in a lattice.

site_indices: a NTuple{2, Int} with indices of sites connected by the bond.

translate_uc: the unit cell offset.

pbc: a vector of boolean values indicating if the bonds should be applied periodically over each axis.

hop_operator: a matrix of type MT representing the operator affecting the internal state.

LatticeModels.bonds — Method

bonds(lattice, hoppings...)

Generates a BondSet for a given set of Hoppings on a given Lattice.

LatticeModels.bonds — Method

bonds(operator)

Generates a BondSet for the provided operator.

LatticeModels.hopping — Function

hopping([hop_operator,] kwargs...)

A convenient constructor for a Hopping object. hop_operator can be either a matrix or a number (in that case a 1×1 matrix will be created automatically)

Keyword arguments:

site_indices: a NTuple{2, Int} (or Int if they are equal) with indices of sites connected by the bond. (1, 1) by default.

translate_uc: the unit cell offset. Zeros by default.

axis: overrides translate_uc and sets its components to zero on all axes except given.

pbc: a vector of boolean values indicating if the bonds should be applied periodically over each axis.

Can also be a single boolean, which will set all elements of the vector to given value. false by default.

If site_indices are equal and translate_uc is zero, this means that the bond connects each site with itself, in which case an error will be thrown. Note that the dimension count for the hopping is dynamic and will automatically change during runtime.

LatticeModels.hopping_operator — Function

hopping_operator(args...)

Creates a hopping operator: $\hat{A} = \sum_{pairs} \hat{t} \hat{c}^\dagger_j \hat{c}_i + h. c.$

hopping_operator(function, lattice, hopping[, field])
hopping_operator(excl_function, lattice, hopping[, field])
hopping_operator(lattice, hopping, excl_value[, field])

Arguments:

excl_function: takes a LatticeSite and its coordinate vector, returns whether the pair with this site should be included.

excl_value: same as excl_function, but represented as a LatticeValue{Bool}

lattice: the lattice to create the operator on.

hopping: the Hopping object describing the site pairs and the $\hat{t}$ operator.

field: the AbstractField object that defines the magnetic field to generate phase factors using Peierls substitution.

LatticeModels.promote_dims! — Method

promote_dims!(hopping, ndims)

Changes dimension count of hopping to ndims if possible.

Magnetic fields

LatticeModels.FluxField — Type

FluxField <: AbstractField

An object representing a small magnetic flux through given point. The field is directed along z-axis. Fields:

B: The magnetic field value

point: A NTuple{2, Number} representing the point where the magnetic flux is located.

LatticeModels.LandauField — Type

LandauField <: AbstractField

An object representing Landau calibrated uniform magnetic field along z-axis. Fields:

B: The magnetic field value

LatticeModels.NoField — Type

NoField <: AbstractField

A stub object representing zero magnetic field. Use it as a default magnetic field argument in functions - this will not cause any performance overhead.

LatticeModels.SymmetricField — Type

SymmetricField <: AbstractField

An object representing symmetrically calibrated uniform magnetic field along z-axis. Fields:

B: The magnetic field value

LatticeModels.apply_field! — Method

apply_field!(hamiltonian, field[; nsteps])

Applies magnetic field to given hamiltonian matrix by adjusting the phase factors.

LatticeModels.trip_integral — Method

trip_integral(field, p1, p2; n_steps=1)

Calculates the $\int_{p1}^{p2} \overrightarrow{A} \cdot \overrightarrow{dl}$ integral using the trapezoidal rule. Increase n_steps to improve accuracy (for linear fields like Landau or symmetrical calibrations the formula is accurate). If needed, redefine this function for specific field types - this is likely to boost accuracy and performance.

LatticeModels.vector_potential — Method

vector_potential(field, point)

Returns vector potential $\overrightarrow{A}$ for field in location point.

This function should be defined for new field types, but it is not necessary unless you want to use built-in trapezoidal rule integrating.

LatticeModels.@field_def — Macro

@field_def block

Defines a new magnetic field type.

Hamiltonian tools

LatticeModels.Spectrum — Type

Spectrum{LT, MT} where {LT<:Lattice, MT<:AbstractMatrix}

Eigenvalues and eigenvectors for some operator.

LatticeModels.filled_projector — Function

filled_projector(spectrum[, fermi_level])

Creates a LatticeOperator that projects onto the eigenvectors which have eigenvalues less than fermi_level (0 by default).

LatticeModels.projector — Method

projector(fun, spectrum)

Creates a LatticeOperator that projects onto the eigenvectors of the spectrum with amplitude defined by the fun functions, which takes the eigenvalue and returns a number (or a boolean).

LatticeModels.projector — Method

projector(spectrum)

Creates a LatticeOperator that projects onto the eigenvectors of the spectrum

LatticeModels.spectrum — Method

spectrum(operator)

Finds eigenvalues and eigenvectors for a LatticeOperator and stores in in a Spectrum.

LatticeModels.@hamiltonian — Macro

@hamiltonian block

Creates a hamiltonian according to the rules defined in the block.

Each line in the block must be a := assignment or a macro-like diagonal/hopping operator description.

The lattice on and the magnetic field for the hamiltonian can be set by assigning lattice and field.

@diag stands for a diagonal part of the hamiltonian - after this you can use a matrix (representing the operator affecting the internal state), a function or a ⊗ tensor product notation.

@hop stands for a hopping part of the hamiltonian - list arguments to pass to the hopping function after this macrocall.

Examples

For example let's generate a Chern insulator hamiltonian.

l = SquareLattice(10, 10)
x, y = coord_values(l)
H = @hamiltonian begin
    lattice := l
    field := LandauField(0.5)   # Landau-calibrated uniform magnetic field, 0.5 flux quanta per 1×1
    @diag (@. abs(x) < 2) ⊗ [1 0; 0 -1]
    @hop [1 im; im -1] / 2 axis = 1
    @hop [1 1; -1 -1] / 2 axis = 2
end

Currents

LatticeModels.AbstractCurrents — Type

AbstractCurrents

Supertype for all type representing currents-like values on a lattice. Subtypes must implement current_lambda and lattice functions.

LatticeModels.DensityCurrents — Type

DensityCurrents <: AbstractCurrents

Density currents for given density matrix and given hamiltonian.

LatticeModels.MaterializedCurrents — Type

MaterializedCurrents <: AbstractCurrents

A AbstractCurrents instance that stores values for all currents explicitly.

LatticeModels.SubCurrents — Type

SubCurrents{CT<:AbstractCurrents} <: AbstractCurrents

A lazy wrapper for a SubCurrents object that representing the same currents but on a smaller lattice.

LatticeModels.current_lambda — Method

current_lambda(::AbstractCurrents)

Returns a function that takes two integer indices of sites in a lattice and returns the current between these two sites.

LatticeModels.lattice — Method

lattice(::AbstractCurrents)

Gets the lattice where the given AbstractCurrents object is defined.

LatticeModels.materialize — Method

materialize([function, ]currents)

Creates a MaterializedCurrents instance for currents.

If function is provided, it must accept a Lattice and two LatticeSites and return if the current between this site must be calculated or not. This can be useful to avoid exsessive calculations.

Unitary evolution

LatticeModels.evolution_operator — Method

evolution_operator(H, t[, k])

Calculates the unitary evolution operator using the formula

$ \mathcal{U}(t) = e^{-\frac{1}{i\hbar} \hat{H} t} $

Arguments

H: the hamiltonian matrix

t: the evolution time

k: if provided, the exponent will be calculated using a Taylor series expansion with order k

LatticeModels.@evolution — Macro

@evolution [kwargs...] {rules...} for_loop

Generates an environment with defined hamiltonian and density matrices that evolve by certain laws. See Unitary evolution for more details.