LatticeModels.jl

LatticeModels.AbstractCurrents

LatticeModels.DensityCurrents

LatticeModels.FluxField

LatticeModels.HoneycombLattice

LatticeModels.LandauField

LatticeModels.Lattice

LatticeModels.LatticeValue

LatticeModels.LatticeValue

LatticeModels.MaterializedCurrents

LatticeModels.NoField

LatticeModels.SiteOffset

LatticeModels.SiteOffset

LatticeModels.SquareLattice

LatticeModels.SubCurrents

LatticeModels.SymmetricField

LatticeModels.TimeSequence

LatticeModels.TimeSequence

LatticeModels.TimeSequence

LatticeModels.check_issublattice

LatticeModels.check_samelattice

LatticeModels.check_samemacrocell

LatticeModels.coord_values

LatticeModels.differentiate!

LatticeModels.evolution_operator!

LatticeModels.haldane

LatticeModels.integrate!

LatticeModels.lattice

LatticeModels.line_integral

LatticeModels.map_currents

LatticeModels.materialize

LatticeModels.pairs_by_distance

LatticeModels.project

LatticeModels.qwz

LatticeModels.radius_vector

LatticeModels.radius_vector

LatticeModels.site_distance

LatticeModels.site_distance

LatticeModels.site_index

LatticeModels.sublattice

LatticeModels.timestamps

LatticeModels.vector_potential

LatticeModels.@evolution

Lattice system

LatticeModels.Lattice — Type

Lattice{LatticeSym, N, NB}

A finite subset of a Brvais{N, NB}. LatticeSym is a Symbol which represents the type of the lattice (e. g. :square, :honeycomb). This makes Lattice object behavior known at compile-time, which allows to introduce various optimizations or to define specific plot recipes.

Lattice(sym, sz, bvs[, mask])

Constructs a finite Lattice{sym, N, NB} as a subset of the bvs Bravais lattice. sz is a NTuple{N, Int} which represents how many times the unit cell of bvs was translated by each axis - these sites form a macrocell. mask, if defined, is a Vector{Bool} storing information about which of the sites from the macrocell are actually included in the lattice, and which are not.

For example, a 3×3 square lattice with its center site excluded is represented as Lattice(:square, (3, 3), Bravais([1 0; 0 1]), Bool[1, 1, 1, 1, 0, 1, 1, 1, 1])

To define a new type of lattice, create an alias for Lattice{YourSym, YourN, YourNB}. Refer to the docs for detailed explanation.

LatticeModels.check_issublattice — Method

Checks if l1 is sublattice of l2. Throws an error if not.

LatticeModels.check_samelattice — Method

Checks if l1 and l2 objects are defined on one lattice. Throws an error if not.

LatticeModels.check_samemacrocell — Method

Checks if l1 and l2 are defined on one macrocell. Throws an error if not.

LatticeModels.radius_vector — Method

radius_vector(l::Lattice, site1::LatticeSite, site2::LatticeSite) -> vector

Finds the vector between two sites on a lattice according to possibly periodic boundary conditions (site2 will be translated along the macrocell to minimize the distance between them).

LatticeModels.site_distance — Method

site_distance(l::Lattice, site1::LatticeSite, site2::LatticeSite[; pbc=false])

Returns the distance between two sites on the l lattice.

Keyword arguments:

pbc: if true, the boundary conditions will be considered periodic and

the distance will be measured on the shortest path.

LatticeModels.site_distance — Method

site_distance(; pbc)

Generates a function that finds the distance between sites (see site_distance(::Lattice, ::LatticeSite, ::LatticeSite)). This notation can be handy when passing this function as an argument.

LatticeModels.site_index — Method

site_index(l::Lattice, site::LatticeSite; macrocell=false)

Returns the integer index for given site in lattice. Returns nothing if the site is not present in the lattice.

LatticeModels.sublattice — Method

sublattice(lf::Function, l::Lattice) -> Lattice

Generates a a subset of lattice l by applying the lf function to its sites. The lf function must return a boolean value.

Lattice values

LatticeModels.LatticeValue — Type

LatticeValue{T, LT}

Represents a value of type T on a Lattice{LatticeSym} lattice.

Fields:

lattice: the Lattice object the value is defined on

values: the values on different sites

LatticeModels.LatticeValue — Method

LatticeValue(l::Lattice, v::AbstractVector)

Constructs a LatticeValue object.

LatticeModels.coord_values — Method

coord_values(l::Lattice)

Generates a tuple of LatticeValues representing spatial coordinates.

LatticeModels.project — Method

project(lv::LatticeValue, axis)

Creates a mapping from site coordinates to values of lv. The coordinate axis to project the sites onto can be set with the axis argument - it can be either an integer from 1 to 3 or an axis descriptor Symbol.

Bonds

LatticeModels.SiteOffset — Type

SiteOffset{T, N}

A struct representing bonds in some direction in a lattice.

SiteOffset([site_indices, ]translate_uc)

Constructs a SiteOffset object.

Arguments:

site_indices: a ::Int => ::Int pair with indices of sites connected by the bond.

When not defined, the resulting bonds object will connect site with any basis index to a site with the same basis index, but in another unit cell.

translate_uc: the unit cell offset.

If site_indices are equal or undefined and translate_uc is zero, the bond connects each site with itself. In this case an error will be thrown. Note that though the dimension count for the bond is static, it is automatically compatible to higher-dimensional lattices.

LatticeModels.SiteOffset — Type

SiteOffset(site_indices)
SiteOffset([site_indices; ]axis[, dist=1])

A convenient constructor for a SiteOffset object. site_indices is 1 => 1 by default.

site_indices is a ::Int => ::Int pair with indices of sites connected by the bond; 1 => 1 is the default value.

Keyword arguments:

axis: The hopping direction axis in terms of unit cell vectors.

dist: The hopping distance in terms of

LatticeModels.radius_vector — Method

radius_vector(l::Lattice, hop::SiteOffset)

Finds the vector between two sites on a lattice according to possibly periodic boundary conditions (site2 will be translated along the macrocell to minimize the distance between them).

Magnetic fields

LatticeModels.NoField — Type

NoField <: AbstractField

A stub object representing zero magnetic field. Use it as a default magnetic field argument in functions - this will not cause any performance overhead.

LatticeModels.line_integral — Method

line_integral(field, p1, p2[, n_steps=1])

Calculates the $\int_{p1}^{p2} \overrightarrow{A} \cdot \overrightarrow{dl}$ integral using the trapezoidal rule. Increase n_steps to improve accuracy (note that for linear fields like Landau or symmetrical calibrations the formula is already pefrectly accurate). If needed, redefine this function for specific field types - this is likely to boost accuracy and performance.

LatticeModels.vector_potential — Method

vector_potential(field, point)

Returns vector potential $\overrightarrow{A}$ for field in location point.

This function should be defined for new field types, but it is not necessary unless you want to use built-in trapezoidal rule integrating.

Operators

Currents

LatticeModels.AbstractCurrents — Type

AbstractCurrents

Supertype for all type representing currents-like values on a lattice. Subtypes must implement Base.getindex(::Int, ::Int) and lattice functions.

LatticeModels.MaterializedCurrents — Type

MaterializedCurrents <: AbstractCurrents

A AbstractCurrents instance that stores values for all currents explicitly.

LatticeModels.SubCurrents — Type

SubCurrents{CT<:AbstractCurrents} <: AbstractCurrents

A lazy wrapper for a Currents object representing the same currents but on a smaller lattice.

LatticeModels.lattice — Method

lattice(::AbstractCurrents)

Gets the lattice where the given AbstractCurrents object is defined.

LatticeModels.map_currents — Method

map_currents(map_fn, currs::AnstractCurrents[; reduce_fn, sort=false])

Accepts a function that takes a Lattice and two LatticeSites and returns any value. Applies map_fn to every site pair and returns two Vectors: one with currents, one with results of map_fn.

Keyword arguments:

reduce_fn: if a function is provided, all currents with the same mapped value will be reduced into one value.

For example, if aggr_fn=(x -> mean(abs.(x))), and map_fn finds the distance between the sites, the returned lists will store the distance between sites and the average absolute current between sites with such distance.

sort: if true, the output arrays will be sorted by mapped value.

LatticeModels.materialize — Method

materialize([function, ]currents)

Creates a MaterializedCurrents instance for currents.

If function is provided, it must accept a Lattice and two LatticeSites and return if the current between this site must be calculated or not. This can be useful to avoid exsessive calculations.

LatticeModels.pairs_by_distance — Method

pairs_by_distance(f)

A selector function used for hopping operator definition or currents materialization.

Takes a function and generates a lambda which accepts a lattice and two LatticeSites, returning whether f applied to distance between the two sites returned true.

Unitary evolution

LatticeModels.evolution_operator! — Method

evolution_operator(H, t)

Calculates the unitary evolution operator using the formula

$ \mathcal{U}(t) = e^{-\frac{1}{i\hbar} \hat{H} t} $

Arguments

H: the hamiltonian matrix

t: the evolution time

LatticeModels.@evolution — Macro

@evolution [kwargs...] {rules...} for_loop

Generates an environment with defined hamiltonian and density matrices that evolve by certain laws. See Unitary evolution for more details.

Keyword arguments:

k: order of the Taylor expansion for matrix exponent. If omitted, the default exp function will be used.

pade: set this to true to use Padé approximant formula instead of Taylor expansion.

rtol: the relative tolerance to decide whether the Δt changed between iterations or not. 1e-12 by default.

show_progress: defines whether the progress informer must be displayed or not. true by default.

Time sequences

LatticeModels.TimeSequence — Type

TimeSequence{ET}

Series of some data depending on time.

LatticeModels.TimeSequence — Method

TimeSequence(times, values)

Constructs a TimeSequence with given timestamps and values.

LatticeModels.TimeSequence — Method

TimeSequence(value; t=0)

Constructs a TimeSequence with one single snapshot. The timestamp is zero by default but can be over riden.

LatticeModels.differentiate! — Method

differentiate!(ts::TimeSequence)

Differentiate the values stored in the TimeSequence object by time using the symmetric difference formula.

LatticeModels.integrate! — Method

integrate(ts::TimeSequence)

Integrates the values stored in the TimeSequence object over time using the trapezoidal rule.

LatticeModels.timestamps — Method

timestamps(ts::TimeSequence)

Returns the timestamps of the snapshots.

Model zoo

LatticeModels.DensityCurrents — Type

DensityCurrents <: AbstractCurrents

Density currents for given density matrix and given hamiltonian.

LatticeModels.FluxField — Type

FluxField <: AbstractField

An object representing a small magnetic flux through given point. The field is directed along z-axis. Fields:

B: The magnetic field value

point: A NTuple{2, Number} representing the point where the magnetic flux is located.

LatticeModels.HoneycombLattice — Type

HoneycombLattice

Type alias for Lattice{:honeycomb,2,2}.

HoneycombLattice(sz::Vararg{Int, 2})

Constructs a honeycomb lattice with a sz-size macrocell.

LatticeModels.LandauField — Type

LandauField <: AbstractField

An object representing Landau calibrated uniform magnetic field along z-axis. Fields:

B: The magnetic field value

LatticeModels.SquareLattice — Type

SquareLattice{N}

Type alias for Lattice{:square,N,1}.

SquareLattice(sz::Int...)

Constructs a square lattice of size sz.

LatticeModels.SymmetricField — Type

SymmetricField <: AbstractField

An object representing symmetrically calibrated uniform magnetic field along z-axis. Fields:

B: The magnetic field value

LatticeModels.haldane — Function

haldane(l::HoneycombLattice, t1::Real, t2::Real[, m::Real=0; field::AbstractField])

\[\hat{H} = \sum_i^\text{sublattice A} m c^\dagger_i c_i + \sum_j^\text{sublattice B} m c^\dagger_j c_j + \sum_{i, j}^\text{adjacent} \left( t_1 c^\dagger_i c_j + h. c. \right) + \sum_{i, j}^\text{2-connected,\\counter-clockwise} \left( i \cdot t_2 c^\dagger_i c_j + h. c. \right)\]

Generates a Haldane topological insulator hamiltonian operator.

LatticeModels.qwz — Method

qwz([f, ]mv::LatticeValue[; field::AbstractField, pbc=false])
qwz([f, ]l::SquareLattice[, m::Number=1; field::AbstractField, pbc=false])

\[\hat{H} = \sum_i^\text{sites} m_i c^\dagger_i \sigma_z c_i + \sum_i^\text{sites} \left( c^\dagger_{i + \hat{x}} \frac{\sigma_z - i \sigma_x}{2} c_i + c^\dagger_{i + \hat{y}} \frac{\sigma_z - i \sigma_y}{2} c_i + h. c. \right)\]

Generates a QWZ model hamiltonian operator with set magnetic field. If the $m_i$ values are set by the mv::LatticeValue, which must be defined on a SquareLattice. Otherwise they will all be set to m.

f here must be a function or a PairSelector describing which hoppings will be excluded.