Models

Ising model

LatticeModels.Ising — Type

Ising

Struct to hold an Ising configuration plus parameters.

Alter only through set_... functions to ensure consistency.

Analytical results

The following functions return exact values (in the thermodynamic limit), using Onsager's results. These are taken from the IsingModels.jl package by Jorge Fernandez-de-Cossio-Diaz.

LatticeModels.Ising_critical_temperature — Constant

Ising_critical_temperature

Critical temperature of the Ising model on the (infinite) square lattice, obtained by Onsager.

LatticeModels.Onsager_magnetization — Function

Onsager_magnetization(β)

Analytical magnetization (per spin) of the Ising model on the square lattice, obtained by Onsager in the thermodynamic limit.

LatticeModels.Onsager_internal_energy — Function

Onsager_internal_energy(β)

Analytical internal energy (per spin) of the Ising model on the square lattice, obtained by Onsager in the thermodynamic limit.

LatticeModels.Onsager_heat_capacity — Function

Onsager_heat_capacity(β)

Specific heat capacity of the Ising model on the square lattice, as obtained by Onsager in the thermodynamic limit.

Gaussian model